+++ imports/api/genericMethods.test.js
@@ -47,18 +47,20 @@ describe('saveProperties Function', () => {
     ).toThrow();
   });
 });
-describe('addNote Function', () => {
+describe('logInteraction Function', () => {
   it('does not throw', () => {
     const collection = new Mongo.Collection();
     collection.docs = [{}];
     Meteor.users.docs = [{ username: 'a' }];
-    expect(() => GM.addNote(collection, 'a', 'my note')).not.toThrow();
+    expect(() =>
+      GM.logInteraction(collection, 'a', 'my note', 'NOTE')
+    ).not.toThrow();
   });
   it('throws if objectId is not a string', () => {
     const collection = new Mongo.Collection();
     collection.docs = [{}];
     Meteor.users.docs = [{ username: 'a' }];
-    expect(() => GM.addNote(collection, 1, 'a')).toThrow();
+    expect(() => GM.logInteraction(collection, 1, 'a', 'NOTE')).toThrow();
   });
 });
 describe('buildGenericMethods Function', () => {
@@ -69,6 +71,9 @@ describe('buildGenericMethods Function', () => {
     expect(() => methods['a.create']({ name: 'a' })).not.toThrow();
     expect(() => methods['a.saveProperties']('a', { name: 'a' })).not.toThrow();
     expect(() => methods['a.addNote']('a', 'b')).not.toThrow();
+    expect(() => methods['a.logCall']('a', 'b')).not.toThrow();
+    expect(() => methods['a.logEmail']('a', 'b')).not.toThrow();
+    expect(() => methods['a.logMeeting']('a', 'b')).not.toThrow();
     expect(() => methods['a.search']('a')).not.toThrow();
   });
 });