+++ imports/components/sections/interactionMenu/InteractionMenuDisplay.test.js
@@ -5,58 +5,63 @@ import InteractionMenuDisplay from './InteractionMenuDisplay';
 
 describe('InteractionMenuDisplay Component', () => {
   let wrapper;
+  const props = { interactions: [] };
   beforeEach(() => {
-    wrapper = shallow(<InteractionMenuDisplay />);
+    wrapper = shallow(<InteractionMenuDisplay {...props} />);
   });
   afterEach(() => {
     wrapper.unmount();
   });
-  it('calls setNote on change of note', () => {
-    const setNote = jest.fn();
-    wrapper.setProps({ setNote });
-    wrapper.find('#note').simulate('change', { target: { value: 'a' } });
-    expect(setNote).toHaveBeenCalled();
+  it('renders without error', () => {});
+  it('renders the various interactions from the props', () => {
+    expect(wrapper.find('.interaction-item').length).toBe(0);
+    wrapper.setProps({
+      interactions: ['NEW_NOTE', 'LOG_CALL', 'LOG_EMAIL', 'LOG_MEETING'],
+    });
+    expect(wrapper.find('.interaction-item').length).toBe(4);
   });
-  it('sets the note button group class to expanded if isWritingNote is true', () => {
-    wrapper.setProps({ isWritingNote: true });
-    expect(wrapper.find('.menu-panel .button-group').hasClass('expanded')).toBe(
-      true
-    );
-    expect(
-      wrapper.find('.menu-panel .button-group').hasClass('expandable')
-    ).toBe(false);
+  it('calls setActiveInteraction on click of an interaction', () => {
+    const setActiveInteraction = jest.fn();
+    wrapper.setProps({
+      interactions: ['NEW_NOTE', 'LOG_CALL', 'LOG_EMAIL', 'LOG_MEETING'],
+      setActiveInteraction,
+    });
+    expect(setActiveInteraction).toHaveBeenCalledTimes(0);
+    wrapper
+      .find('button')
+      .at(0)
+      .simulate('click');
+    expect(setActiveInteraction).toHaveBeenCalledTimes(1);
+    wrapper
+      .find('button')
+      .at(1)
+      .simulate('click');
+    expect(setActiveInteraction).toHaveBeenCalledTimes(2);
+    wrapper
+      .find('button')
+      .at(2)
+      .simulate('click');
+    expect(setActiveInteraction).toHaveBeenCalledTimes(3);
+    wrapper
+      .find('button')
+      .at(3)
+      .simulate('click');
+    expect(setActiveInteraction).toHaveBeenCalledTimes(4);
   });
-  it('sets the note button group class to expandable if isWritingNote is false', () => {
-    wrapper.setProps({ isWritingNote: false });
+  it('ignores invalid interactions', () => {
+    wrapper.setProps({ interactions: ['a', 'b'] });
+    expect(wrapper.find('button').length).toBe(0);
+  });
+  it('sets the active class when the interaction matches', () => {
+    wrapper.setProps({
+      interactions: ['NEW_NOTE', 'LOG_CALL'],
+      activeInteraction: 'LOG_CALL',
+    });
     expect(
-      wrapper.find('.menu-panel .button-group').hasClass('expandable')
+      wrapper
+        .find('button')
+        .at(1)
+        .hasClass('active')
     ).toBe(true);
-    expect(wrapper.find('.menu-panel .button-group').hasClass('expanded')).toBe(
-      false
-    );
-  });
-  it('calls addNote on click of save if isWritingNote is true', () => {
-    const addNote = jest.fn();
-    wrapper.setProps({ addNote, isWritingNote: true });
-    wrapper.find('.menu-panel .button-primary').simulate('click');
-    expect(addNote).toHaveBeenCalled();
-  });
-  it('does not call addNote on click of save if isWritingNote is false', () => {
-    const addNote = jest.fn();
-    wrapper.setProps({ addNote, isWritingNote: false });
-    wrapper.find('.menu-panel .button-primary').simulate('click');
-    expect(addNote).not.toHaveBeenCalled();
-  });
-  it('calls cancelNote on click of save if isWritingNote is true', () => {
-    const cancelNote = jest.fn();
-    wrapper.setProps({ cancelNote, isWritingNote: true });
-    wrapper.find('.menu-panel .button-secondary').simulate('click');
-    expect(cancelNote).toHaveBeenCalled();
-  });
-  it('does not call cancelNote on click of save if isWritingNote is false', () => {
-    const cancelNote = jest.fn();
-    wrapper.setProps({ cancelNote, isWritingNote: false });
-    wrapper.find('.menu-panel .button-secondary').simulate('click');
-    expect(cancelNote).not.toHaveBeenCalled();
   });
 });